requirements.txt
pydantic>=2.5.0
python-dotenv>=1.0.0
google-generativeai>=0.7.0  # Библиотека для работы с Gemini
colorama>=0.4.6
networkx>=3.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
numpy>=1.24.0
tenacity>=8.5.0
python-louvain>=0.16
langchain-text-splitters
